Add REPL history to .gitignore & add ch6 project
Minor addition to the top-level '.gitignore' file; new ignore-patterns
are, '**/.nrepl-history' and '**/.lein-repl-history' to prevent
leiningen's and boot's REPL history files from being added to version
control. (This also marks the removal of 'peg-game/.lein-repl-history',
which was previously tracked, from version control).

Create the skeleton for the ch6 leiningen project, which is named
'the-divine-cheese-code/'. This includes the 'src/' hierarchy as well as
a very short 'README.md' file stating that the project directory is
associated with Chapter 6 of CFTBAT.